re-organize rpc-related examples, add example server for statistics
