+++ examples/grpc_vpp/remote_client/README.md
@@ -1,0 +1,7 @@
+
+How to run example:
+1. Run GRPC server: [vpp-agent-grpc](../../cmd/vpp-agent-grpc)
+2. Run GRPC client:
+```
+go run main.go
+```
