+++ examples/grpc_vpp/remote_client/doc.go
@@ -1,0 +1,3 @@
+// Example remote_client demonstrates remote VPP config management
+// via clientv1 using GPRC transport.
+package main
