+++ examples/grpc_vpp/statistics/doc.go
@@ -1,0 +1,2 @@
+// Example statistics app demonstrates usage of grpc to stream vpp statistics.
+package main
