--- examples/remoteclient_grpc_vpp/README.md
@@ -1,7 +1,0 @@
-
-How to run example:
-1. Run GRPC server: [vpp-agent-grpc](../../cmd/vpp-agent-grpc)
-2. Run GRPC client:
-```
-go run main.go
-```
