--- examples/remoteclient_grpc_vpp/doc.go
@@ -1,3 +1,0 @@
-// Example remoteclient_vpp demonstrates remote VPP config management
-// via clientv1 (using GPRC transport).
-package main
